% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{applyGbmModels}
\alias{applyGbmModels}
\title{applyGbmModels}
\usage{
applyGbmModels(models, data, preds)
}
\arguments{
\item{models}{A list of GBM models created during the training phase.}

\item{data}{The test data.}

\item{preds}{A data frame containing the predictions of models that have already been run.}
}
\description{
Apply the GBM models to the test data and augment the predictions created thus
far with the GBM models' predictions.
}

