% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{applyMaBoostModels}
\alias{applyMaBoostModels}
\title{applyMaBoostModels}
\usage{
applyMaBoostModels(models, data, preds)
}
\arguments{
\item{models}{A list of MaBoost models created during the training phase.}

\item{data}{The test data.}

\item{preds}{A data frame containing the predictions of models that have already been run.}
}
\description{
Apply the MaBoost models to the test data and augment the predictions created thus
far with the MaBoost models' predictions.
}

