% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{applyPrimaryRfModel}
\alias{applyPrimaryRfModel}
\title{applyPrimaryRfModel}
\usage{
applyPrimaryRfModel(model, data)
}
\arguments{
\item{model}{The primary Random Forest model.}

\item{data}{The test data.}
}
\description{
Generate productions from the primary Random Forest model.
}

