% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{buildMetaModels}
\alias{buildMetaModels}
\title{buildMetaModels}
\usage{
buildMetaModels(numSplits, data)
}
\arguments{
\item{numSplits}{The number of splits used to generate predictions.}

\item{data}{The training data.}
}
\description{
buildMetaModels
}

