% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{loadAllTrainingData}
\alias{loadAllTrainingData}
\title{loadAllTrainingData}
\usage{
loadAllTrainingData(numSplits, setAside, loadFromCache = FALSE)
}
\arguments{
\item{loadFromCache}{TRUE if we should attempt to load the output of this function from the cache.}
}
\description{
Load and prepare the training data.  Optionally set aside a portion for final
cross-validation. Add the 'cvSet' column to uniquely identify sets of the training
(non-set-aside) data for internal iterations of smaller sets of cross-validation.
}

