% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{loadDataForSubmission}
\alias{loadDataForSubmission}
\title{loadDataForSubmission}
\usage{
loadDataForSubmission(loadFromCache = FALSE)
}
\arguments{
\item{loadFromCache}{TRUE if we should attempt to load the output of this function from the cache.}
}
\description{
Load the test set and prepare it for submission.
}

