% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_function.R
\name{onp.predict}
\alias{onp.predict}
\title{onp.predict}
\usage{
onp.predict(test, train = NULL, verbose = TRUE, store = FALSE,
  dest.folder = NULL, visible = TRUE, ntree = NA, nodesize = NA,
  seed = NA, ...)
}
\arguments{
\item{test}{A data frame containing a test set with 61 features.}

\item{train}{An optional data frame containing a training set with 62 features.  If not provided, a default internal training set is used.}

\item{verbose}{TRUE if the function should print progress.}

\item{store}{TRUE if the resulting predictions should be saved to a CSV file.}

\item{dest.folder}{Folder in which the results will be saved.}

\item{visible}{TRUE if the results should be visible when returned.}

\item{ntree}{Number of trees in the random forest.}

\item{nodesize}{Ending node size for the random forest.}

\item{seed}{Starting random seed for the random forest.}

\item{...}{Additional parameters to the model generator randomForest().}
}
\value{
A dataframe containing two columns: id, popularity
}
\description{
Use features, outlier detection, and hyperparameter optimization brought to
you by the Unlikelies to predict online news popularity.
}
\details{
The Unlikelies (c) 2016.03.17
Authors: Roger Cusco, Matthew Sudmann-Day, Miquel Torrens

DISCLAIMER: Evaluator's eyes only. This function contains confidential
procedures for the Kaggle competition results and must not be released or
displayed to third parties under any circumstance without express written
permission by the authors.
}
\examples{
# predictions <- onp.predict(test)
# predictions <- onp.predict(test,train)

}

