% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{produceFinalPredictions}
\alias{produceFinalPredictions}
\title{produceFinalPredictions}
\usage{
produceFinalPredictions(numSplits, data)
}
\arguments{
\item{numSplits}{The number of splits used to generate predictions.}

\item{data}{The test data.}
}
\description{
Generate final predictions based on the original test data and the output of
all other models applied to that test data.
}

