% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{splitForCrossValidation}
\alias{splitForCrossValidation}
\title{splitForCrossValidation}
\usage{
splitForCrossValidation(data, index)
}
\arguments{
\item{data}{The complete non-set-aside data set.}

\item{index}{The index of the split to use.}
}
\description{
Split the outer training (non-set-aside) data into a smaller training and test set.
This split is done based on the previously created cvSet column.  Calling this
function with all values of index (from 1 to the number of splits) ensure all
combinations are used.
}

