% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{trainMetaModel}
\alias{trainMetaModel}
\title{trainMetaModel}
\usage{
trainMetaModel(preds, key, loadFromCache = FALSE)
}
\arguments{
\item{preds}{The predictions of all previous models in combination with the original data.}

\item{key}{A key to uniquely identify the results of this function in the cache.}

\item{loadFromCache}{TRUE if we should attempt to load the output of this function from the cache.}
}
\description{
Train a "meta" model, a Random Forest model, which synthesizes the original training data
in combination with the output of all other models.
}

