% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{trainOneGbmModel}
\alias{trainOneGbmModel}
\title{trainOneGbmModel}
\usage{
trainOneGbmModel(data)
}
\arguments{
\item{data}{The training data.}
}
\description{
Train a single GBM model.  This function exists to serve trainMultipleGbmModels().
}

