% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{trainPrimaryRfModel}
\alias{trainPrimaryRfModel}
\title{trainPrimaryRfModel}
\usage{
trainPrimaryRfModel(data, key, loadFromCache = FALSE)
}
\arguments{
\item{data}{The training data.}

\item{key}{A key to identify this batch of training data for caching purposes.}

\item{loadFromCache}{TRUE if we should attempt to load the output of this function from the cache.}
}
\description{
Train the primary Random Forest model from which all other models are built.
}

